# swagger.yaml
swagger: '2.0'
info:
  title: gas
  version: 1.0.0
host: 'vvdev.ru:1334'
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
...
  /curators/login:
    post:
      tags:
      - Curator
      summary: Вход
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
      responses:
        200:
          description: Токен
  /curators/users:
    post:
      tags:
      - Curator
      summary: создать пользователя
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            firstName:
              type: string
            lastName: 
              type: string
            middleName:
              type: string
            password:
              type: string
            repeatPassword:
              type: string
            email:
              type: string
            position:
              type: string
            cons_inn:
              type: string
            cons_kpp:
              type: string
      responses:
        200:
          description: ok
  /curators/notifications:
    post:
      tags:
      - Curator
      summary: рассылка 
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            userIds:
              type: array
              items:
                type: string
            cons_UIDs: 
              type: array
              items:
                type: string
            type:
              type: string
            text:
              type: string
            allUsers:
              type: boolean
            allConsumers:
              type: boolean
      responses:
        200:
          description: ok
  /curators/consumers:
    get:
      tags:
      - Curator
      summary: Список cons_UID для куратора
      parameters:
      - in: query
        name: email
        type: string
        description: Почту прокидывает админ если хочет получить список консамеров куратора
      - in: query
        name: cons_inn
        type: string
      - in: query
        name: cons_short_name
        type: string
      - in: query
        name: user_name
        type: string
      - in: query
        name: sortBy
        type: string
      - in: query
        name: desc
        type: boolean
      - in: query
        name: cons_UIDs
        type: string
        description: array
      - in: query
        name: lkp
        type: string
      - in: query
        name: cons_kpp
        type: string
      responses:
        200:
          description: ок
...
  /curators/many:
    put:
      tags:
      - Curator
      summary: Обновить список кураторов
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            ids:
              type: string
            blocked:
              type: boolean
            tabs:
              type: string
      responses:
        200:
          description: ок
    delete:
      tags:
      - Curator
      summary: Удалить список кураторов
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            ids:
              type: string
      responses:
        200:
          description: ок
  /curators:
    get:
      tags:
      - Curator
      summary: Список кураторов для админа лк
      parameters:
      - in: query
        name: position
        type: string
      - in: query
        name: sort
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: email
        type: string
      responses:
        200:
          description: ок
    post:
      tags:
      - Curator
      summary: Создать куратора 
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
            repeatPassword:
              type: string
            email:
              type: string
            phone:
              type: string
            position:
              type: string
            blocked:
              type: boolean
            firstName:
              type: string
            lastName:
              type: string
            middleName:
              type: string
      responses:
        200:
          description: ок
  /curators/{id}:
    get:
      tags:
      - Curator
      summary: Получить куратора 
      parameters:
      - in: path
        name: id
        required: true
        type: string
      responses:
        200:
          description: ок
    delete:
      tags:
      - Curator
      summary: удалить куратора 
      parameters:
      - in: path
        name: id
        required: true
        type: string
      responses:
        200:
          description: ок
    put:
      tags:
      - Curator
      summary: Обновить куратора 
      parameters:
      - in: path
        name: id
        required: true
        type: string
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
            repeatPassword:
              type: string
            email:
              type: string
            phone:
              type: string
            position:
              type: string
            blocked:
              type: boolean
      responses:
        200:
          description: ок
...
  /users/login:
    post:
      tags:
      - User
      summary: Вход
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
      responses:
        200:
          description: Токен
          schema:
            type: object
            properties:
              token:
                type: string
              _id:
                type: string
                description: id пользователя
              consumers:
                  type: array
                  items:
                    type: object
                    properties:
                      login:
                        type: string
                      cons_UID:
                        type: string
              cons_UIDs:
                type: array
                items:
                  type: string
        401:
          description: Пароль неправильный
        500:
          description: Ошибка сервера
  /users/{id}:
    delete:
      tags:
      - User
      summary: удалить пользователя(будут удалены все консамеры и сублогины относящиеся к пользователю)
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
...
  /locations:
    get:
      tags:
      - Locations
      summary: Список площадок + список mets для каждой площадки + список equipments для каждого mets (доступные каждому юзеру)
      parameters:
      - in: query
        name: contr_UID
        type: string
      - in: query
        name: work
        type: boolean
      - in: query
        name: equipment_type
        type: string
      responses:
        200:
          description: Список площадок
          schema:
            $ref: '#/definitions/Location'
        500:
          description: Ошибка сервера
...
  /locations/{id}/mets:
    get:
      tags:
      - Locations
      parameters:
      - in: query
        name: work
        type: boolean
      - in: query
        name: equipment_type
        type: string
      - in: path
        name: id
        type: string
        required: true
      - in: query
        name: st_date
        type: string
      - in: query
        name: end_date
        type: string
      responses:
        200:
          description: ok
  /equipments/{id}:
    put:
      summary: Ввод показаний
      tags:
      - Equipments
      parameters:
      - in: path
        name: id
        required: true
        type: string
      - in: body
        name: body
        schema:
          type: object
          properties:
            location_uuid:
              type: string
            Met_uuid: 
              type: string
            Dt_value:
              type: string
            Value:
              type: number
            values:
              type: array
              items:
                type: object
                properties:
                  Dt_value:
                    type: string
                    description: Дата(день)
                  Value:
                    type: number
            valueType:
              type: number
              description: 1-ввод показаний 2-ввод обьемов
            contr_UID:
              type: string
            GasVolume:
              type: number
      responses:
        200:
          description: OK
  /equipments/import/{fileName}:
    post:
      tags:
      - Equipments
      summary: Entering values from excel file
      parameters:
      - in: path
        name: fileName
        required: true
        type: string
        description: only .xlsx/.csv files
      - in: body
        name: body
        schema:
          type: object
          properties:
            valueType:
              type: number
            equipment_uuid:
              type: string  
      responses:
        200:
          description: ok
        400:
          description: Enter all values
        500:
          description: server error
  /equipments/confirm:
    put:
      tags:
      - Equipments
      summary: Отправить файл подтверждающий показания
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            fileName:
              type: string
            location_uuid:
              type: string
            Met_uuid:
              type: string
      responses:
        200:
          description: ok
        500:
          description: server error
  /equipments/export:
    get:
      tags:
      - Equipments
      summary: export data in xlsx file
      parameters:
      - in: query
        name: contr_UID
        type: string
      - in: query
        name: location_uuid
        type: string
      - in: query
        name: Met_uid
        type: string
      - in: query
        name: valueType
        type: number
      responses:
        200:
          description: file name
          schema:
            type: object
            properties:
              fileName:
                type: string
  /equipments/{id}/values:
    get:
      tags:
      - Equipments
      summary: Получить показания
      parameters:
      - in: path
        name: id
        required: true
        type: string
      responses:
        200:
          description: OK
  /equipments/gas_equipment_info:
    get:
      tags:
      - Equipments
      parameters:
      - in: query
        name: work
        type: boolean
      - in: query
        name: equipment_type
        type: string
      - in: query
        name: location_uuids
        type: array
        items:
          type: string
      responses:
        200:
          description: OK
        404:
          description: resource not found
        500:
          description: server error
  /equipments/qGetScanEquipmentValues:
    get:
      tags:
      - Equipments
      parameters:
      - in: query
        name: file_id
        type: string
      responses:
        200:
          description: OK
  /equipments/qDelScanEquipmentValues:
    delete:
      tags:
      - Equipments
      parameters:
      - in: query
        name: file_id
        type: string
      responses:
        200:
          description: OK
  /equipments/qSetScanEquipmentValues:
    put:
      tags:
      - Equipments
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            location_uuid:
              type: string
            met_uuid:
              type: string
            dt_value:
              type: string
            name:
              type: string
            fileName:
              type: string
            equipment_uuid:
              type: string
      responses:
        200:
          description: OK
...
  /users/consumers:
    post:
      tags:
      - User
      summary: Добавить консамера
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            email:
              type: string
            cons_inn:
              type: string
            cons_kpp:
              type: string
      responses:
        200:
          description: OK
        500:
          description: Ошибка сервера
...
definitions:
...
  Location:
    type: object
    properties:
      location_uuid:
        type: string
      location_Name:
        type: string
      location_address:
        type: string
      met_units:
        type: array
        items:
          $ref: '#/definitions/Met'
  Met:
    type: object
    properties:
      Met_uuid:
        type: string
      Met_name:
        type: string
      equipments:
        $ref: '#/definitions/Equipment'
  Equipment:
    type: object
    properties:
      equipment_uuid:
        type: string
      Equipment_type:
        type: string
      equipment_name: 
        type: string
      Equipment_max_power:
        type: string
      serial_number:
        type: string
      Work:
        type: boolean
      Values:
        type: array
        items:
          type: object
          properties:
            Dt_value:
              type: string
            Value:
              type: string
            basic:
              type: string
...